Move functions from common to ProfileLoader
