+++ src/AvTranscoder/ProfileLoader.hpp
@@ -73,5 +73,10 @@ private:
 	Profiles _profiles;
 };
 
+#ifndef SWIG
+void split( std::vector< std::string >& splitedString, const std::string& inputString, const std::string& splitChars = ";" );
+int getFilesInDir( const std::string& dir, std::vector< std::string >& files );
+#endif
+
 }
 #endif
